Develop

+++ src/Application/UseCases/Comment/DeleteComment.ts
@@ -3,13 +3,15 @@ import { Comment } from "../../../Domain/Entities/Comment";
 import { inject, injectable } from "tsyringe";
 
 @injectable()
-export class DeleteComment {
+export class DeleteComment
+{
   constructor(
     @inject("CommentRepository") private commentRepository: CommentRepository,
-  ) {}
+  ) { }
 
-  async execute(id: string): Promise<{ message: string }> {
-    const { message } = await this.commentRepository.deleteComment(id);
+  async execute(id: string, userId: string): Promise<{ message: string }>
+  {
+    const { message } = await this.commentRepository.deleteComment(id, userId);
 
     return { message };
   }
